Feature/implement token authentication

+++ Application/Dtos/Responses/AccessHistoryResponse.cs
@@ -2,9 +2,9 @@
 {
     public class AccessHistoryResponse
     {
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public string Username { get; set; } = null!;
-        public int DoorId { get; set; }
+        public long DoorId { get; set; }
         public string Doorname { get; set; } = null!;
         public bool AccessGranted { get; set; }
         public DateTime Timestamp { get; set; }
